Fix all clippy warnings
- gc/allocator.rs: Use saturating_sub, allow mut_from_ref for unsafe GC code
- gc/collector.rs: Collapse nested if using let chains
- value.rs: Simplify integer range check
- vm/interpreter.rs:
  - Add Default impl for ObjectInstance
  - Implement Iterator trait for ForInIterator and ForOfIterator
  - Replace unwrap_or("") and unwrap_or(Value::undefined()) with unwrap_or_default()
  - Rename to_string_value to stringify_value (naming convention)
  - Merge identical if-else branches
  - Remove redundant wildcard pattern
  - Use saturating_sub
- parser/compiler.rs: Use clamp() instead of max().min()
- parser/lexer.rs: Use is_some_and instead of map_or
- runtime/array.rs: Use Iterator::find instead of manual loop
- runtime/function.rs: Remove unnecessary cast
- runtime/object.rs: Add type annotations to transmute
- util/dtoa.rs: Use Range::contains
- util/mod.rs: Use iterator instead of index loop

+++ src/runtime/function.rs
@@ -240,7 +240,7 @@ impl FunctionBytecode {
     pub fn calculate_stack_size(&mut self) {
         // This is a simplified version - a real implementation
         // would analyze the bytecode to find max stack depth
-        self.stack_size = (self.local_count as u16).saturating_add(16);
+        self.stack_size = self.local_count.saturating_add(16);
     }
 
     /// Serialize to bytes for bytecode file format
